requirements.txt
PyPDF2==3.0.1
sentence-transformers==2.5.1
pgvector==0.2.5
psycopg2-binary==2.9.9
boto3==1.34.58